% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.lambda.R
\name{calc.lambda}
\alias{calc.lambda}
\title{calc.lambda
Calculate lambda that is used as the tuning parameter in the objective function.}
\usage{
calc.lambda(X_count)
}
\arguments{
\item{X_count}{Combination of berved single cell HiC matrix, with each column being the upper triangular of a single cell HiC matrix.}
}
\value{
A number that is used as the tuning parameter in the objective function.
}
\description{
calc.lambda
Calculate lambda that is used as the tuning parameter in the objective function.
}
\examples{
calc.lambda(X_count, percent)
}
