% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.size.factor.R
\name{calc.size.factor}
\alias{calc.size.factor}
\title{calc.size.factor
Calculate the size factor}
\usage{
calc.size.factor(x)
}
\arguments{
\item{x}{Combination of berved single cell HiC matrix, with each column being the upper triangular of a single cell HiC matrix.}
}
\value{
A list of size factor, defined as the relative sequencing depth to the median sequencing depth among all single cells.
}
\description{
calc.size.factor
Calculate the size factor
}
\examples{
calc.size.factor(X_count)
}
