% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmix.R
\name{dmix}
\alias{dmix}
\title{dmix
Estimate probability of SZ.}
\usage{
dmix(x, pars)
}
\arguments{
\item{x}{A vector of data.}

\item{pars}{The estimated Gaussian mixture parameters.}
}
\value{

}
\description{
dmix
Estimate probability of SZ.
}
\examples{
dat=rnorm(100)
paramt=get_mix(dat, nmix=3)
dmix(dat,paramt)
}
