% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_single.R
\name{generate_single}
\alias{generate_single}
\title{Generate_single}
\usage{
generate_single(
  data = str1,
  alpha_0,
  alpha_1,
  beta_l,
  beta_g,
  beta_m,
  alpha,
  n_single
)
}
\arguments{
\item{data}{3D coordinates of single cell.}

\item{alpha_0}{parameter that controls sequence depth of data.}

\item{alpha_1}{parameter that controls sequence depth of data.}

\item{beta_l}{parameter that controls effect size of covariate.}

\item{beta_g}{parameter that controls effect size of covariate.}

\item{beta_m}{parameter that controls effect size of covariate.}

\item{alpha}{Proportion of random true 0.}

\item{n_single}{Number of single cells to be generated.}
}
\value{
A list of underline true count, SZ positions, and generated single cells.
}
\description{
This function simulates single cells from 3D structure.
}
\examples{
#Load 3d structure generated from SIMBA package
load("simba_3strs.rdata")
Set random seed
set.seed(1234)
#Generate 100 random type1 single cells
simudat <- generate_single(data=str1, alpha_0=5.6,alpha_1=-1, beta_l=0.9,beta_g=0.9,beta_m=0.9, alpha=0.2, n_single=10)
}
