% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mix.R
\name{get_mix}
\alias{get_mix}
\title{get_mix
Estimate parameters in the Gaussian mixture distribution.}
\usage{
get_mix(xdata, nmix)
}
\arguments{
\item{xdata}{A vector of data.}

\item{nmix}{Number of Gaussian mixtures.}
}
\value{

}
\description{
get_mix
Estimate parameters in the Gaussian mixture distribution.
}
\examples{
get_mix(rnorm(100), nmix=3)

}
