% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GM_psz.R
\name{GM_Psz}
\alias{GM_Psz}
\title{GM_psz
Calculate Psz of a imputed matrix through a Gaussian mixture mdoel.}
\usage{
GM_Psz(observed, imputed, nmix)
}
\arguments{
\item{observed}{The observed matrix.}

\item{imputed}{A imputed matrix.}

\item{nmix}{Number of Gaussian mixtures.}
}
\value{

}
\description{
GM_psz
Calculate Psz of a imputed matrix through a Gaussian mixture mdoel.
}
\examples{
GM_Psz(simudat_res)
}
