% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm.R
\name{hm}
\alias{hm}
\title{hm}
\usage{
hm(datvec, n)
}
\arguments{
\item{datvec}{A vector of upper triangular mamtrix.}

\item{n}{Number of bins.}
}
\value{
Heatmap of the matrix.
}
\description{
This function draws heatmap of HiC data so that we can visually compares the imputation results.
}
\examples{
# Heatmap of the first single cell in the simulated data.
hm(simudat[,1], 61)

}
