% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_lasso_regression.R
\name{keras_lasso_regression}
\alias{keras_lasso_regression}
\title{keras_lasso_regression
Build the frame work of lasso regression using keras.}
\usage{
keras_lasso_regression(
  Y,
  X,
  epochs = 100,
  batch_size = 128,
  lambda1 = 1,
  lambda2 = 1e+10,
  learning_rate = 1e-04,
  lasso_threshold = 0,
  verbose = TRUE
)
}
\arguments{
\item{Y}{Y}

\item{X}{X}

\item{epochs}{The number of the entire training set going through the entire network.}

\item{batch_size}{The number of examples that are fed to the algorithm at a time.}

\item{lambda1}{Tunning parameter to facilitate feature selection and regularization.}

\item{lambda2}{Tunning parameter to penalize the diagonal element of the parameter to eliminate the trivial solution of representing an expression level as a linear combination of itself.}

\item{learning_rate}{A hyper parameter that controls the speed of adjusting the weights of the network with respect to the loss gradient.}

\item{lasso_threshold}{The threshold that is used as a convergence condition of the objective function.}

\item{verbose}{Whether to output the value of metrics at the end of each epoch. Default is TRUE.}
}
\value{

}
\description{
keras_lasso_regression
Build the frame work of lasso regression using keras.
}
\examples{
keras_lasso_regression(Y, Z, epochs = epochs, batch_size = batch_size, lambda1 = lambda1, lambda2 = lambda2,learning_rate = learning_rate, verbose = verbose)
}
