% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_cluster.R
\name{kmeans_cluster}
\alias{kmeans_cluster}
\title{kmeans_cluster}
\usage{
kmeans_cluster(
  data,
  centers,
  nstart = 50,
  iter.max = 200,
  seed = 1234,
  algorithm = c("Hartigan-Wong", "Lloyd", "Forgy", "MacQueen"),
  trace = FALSE
)
}
\arguments{
\item{data}{The observed or imputed matirx, with each column being the uppertriangular of a single cell HiC matrix.}

\item{centers}{Either the number of clusters, say k, or a set of initial (distinct) cluster centres. If a number, a random set of (distinct) rows in x is chosen as the initial centres.}

\item{nstart}{If centers is a number, how many random sets should be chosen?}

\item{iter.max}{The maximum number of iterations allowed.}

\item{seed}{Random seed.}

\item{algorithm}{Character: may be abbreviated. Note that "Lloyd" and "Forgy" are alternative names for one algorithm.}

\item{trace}{Logical or integer number, currently only used in the default method ("Hartigan-Wong"): if positive (or true), tracing information on the progress of the algorithm is produced. Higher values may produce more tracing information.}
}
\value{
Kmeans clustering results.
}
\description{
kmeans_cluster
}
\examples{
kmeans_cluster(data=simudat, centers=2, seed=1250)
}
