% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnorm.R
\name{lnorm}
\alias{lnorm}
\title{lnorm
Normalize and log transform data with 1.01 pseudocount. Used as the input of Gaussian mixture model.}
\usage{
lnorm(raw_count)
}
\arguments{
\item{raw_count}{}
}
\value{

}
\description{
lnorm
Normalize and log transform data with 1.01 pseudocount. Used as the input of Gaussian mixture model.
}
