% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_normalization.R
\name{log_normalization}
\alias{log_normalization}
\title{log_normalization
Preprocess the data by normalizing the data so that each single cell have the same sequencing depth as the median sequencing depth and take a log transformation with pseudocount 1.}
\usage{
log_normalization(x)
}
\arguments{
\item{x}{Combination of berved single cell HiC matrix, with each column being the upper triangular of a single cell HiC matrix.}
}
\value{
A matrix after normalization and log transformation.
}
\description{
Take log normalization of the data.
}
\examples{
log_normalization(X_count)
}
