% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{objective}
\alias{objective}
\title{objective
Define the objective function.}
\usage{
objective(X, lambda, A, B)
}
\arguments{
\item{X}{Combination of berved single cell HiC matrix, with each column being the upper triangular of a single cell HiC matrix.}

\item{lambda}{The tunning parameter of objective function.}

\item{A}{Coefficient matrix A}

\item{B}{Coefficient matrix B}
}
\value{

}
\description{
objective
Define the objective function.
}
\examples{
objective(X, lambda1, A, B)
}
