% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizer_define.R
\name{regularizer_define}
\alias{regularizer_define}
\title{regularizer_define
Define regularizer}
\usage{
regularizer_define(weight_matrix, lambda1 = 1, lambda2 = 1e+10)
}
\arguments{
\item{weight_matrix}{Weighte matrix}

\item{lambda1}{Lambda1}

\item{lambda2}{Lambda2}
}
\value{

}
\description{
regularizer_define
Define regularizer
}
\examples{
regularizer_define(weight_matrix = weight_matrix ,lambda1 = lambda1,  lambda2 = lambda2)
}
