% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRS_ROC.R
\name{SRS_ROC}
\alias{SRS_ROC}
\title{SRS_ROC}
\usage{
SRS_ROC(single, truecount, result)
}
\arguments{
\item{single}{Observed sngle cell with each column being the upper triangular of single cell.}

\item{truecount}{Underline true count of simulated data.}

\item{result}{Result from SRS funciton.}
}
\value{
A plot of ROC curve.
}
\description{
This package draws ROC (Receiver operating characteristic) curve to visually demonstrate ability
to tell SZ from DO.
}
\examples{
SRS_ROC(simudat,simudat_true,simudat_res)
}
