% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summa.R
\name{summa}
\alias{summa}
\title{summa}
\usage{
summa(truecount, predicted, observed, prob, cut = 0.5)
}
\arguments{
\item{truecount}{Underline true count of the simulated data.}

\item{single}{Observed single cells matrix, with each column being the upper triangular of a single cell.}

\item{imputeduted}{Imputed matrix from SRS.}
}
\value{
A list of accuracy measurements with mean and standard error.
}
\description{
This function summarizes the accuracy of simulated data.
}
\examples{
summa(data$singledat, data$truecount,res$imputed)
}
