% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summa2.R
\name{summa2}
\alias{summa2}
\title{summa2}
\usage{
summa2(single, truecount, result)
}
\arguments{
\item{single}{Observed single cells matrix with each column being the upper triangular of a single cell.}

\item{truecount}{Underline true counts from simulation.}

\item{result}{Result form SRS function.}
}
\value{
A vector of PTDO and its SD when fixing PTSZ to be 0.95, and the threshold used in that case.
}
\description{
This function calculates PTDO when fix PTSZ=0.95.
}
\examples{
summa2(data$singledat, data$truecount, res)
}
