% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne_plot.R
\name{tsne_plot}
\alias{tsne_plot}
\title{tsne_plot}
\usage{
tsne_plot(
  data,
  cell_type,
  dims = 2,
  perplexity = 10,
  check_duplicates = FALSE,
  seed = 1234,
  kmeans = TRUE,
  ncenters
)
}
\arguments{
\item{data}{The observed matrix, with each column being the uppertriangular of a single cell HiC matrix.}

\item{cell_type}{A vector that indicates cell type.}

\item{dims}{Integer. Output dimentionality. Default=2.}

\item{perplexity}{Numeric; Perplexity parameter (should not be bigger than 3 * perplexity < nrow(X) - 1, see details for interpretation).}

\item{check_duplicates}{Logical; Checks whether duplicates are present. It is best to make sure there are no duplicates present and set this option to FALSE, especially for large datasets (default: TRUE).}

\item{seed}{Random seed.}

\item{kmeans}{Logical. Whether apply kmeans clustering on tsne data or not.}

\item{ncenters}{Number of clusters in kmeans clustering.}
}
\value{
A stne visualization plot.
}
\description{
tsne_plot
}
\examples{
tsne_plot(simudat, cell_type=c(rep("L4", 131), rep("L5",180)), seed=1250)
}
