% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRS.R
\name{SRS}
\alias{SRS}
\title{SRS
Self representation smoothing of single cell Hi-C matrix.}
\usage{
SRS(
  scHiC,
  expected,
  windowsize = 2,
  nbins,
  lambda1 = NULL,
  lambda2 = 1e+10,
  initA = NULL,
  initS = NULL,
  ncores = 1,
  MAX_ITER = 4,
  ABSTOL = 0.001,
  learning_rate = 1e-04,
  epochs = 100,
  batch_size = 128,
  run_batch = TRUE,
  verbose = TRUE,
  estimates.only = FALSE
)
}
\arguments{
\item{scHiC}{The single-cell Hi-C matrix. It can take three types of formats. The preferred format is a single-cell matrix with
each column being a vector of the upper triangular matrix without including the diagonal entries
of the 2D matrix of a single-cell. Another types of formats are a list with each element being a
2D single-cell contact matrix, or a 3D (\eqn{n\times n\times k}) array that has k matrices of dimension
\eqn{n\times n}. scHiCSRS automatically transforms these two types of input into a matrix with each
column being the vector of upper triangular matrix of a single-cell. For a single-cell matrix of
size \eqn{n \times n}, the length of the vector should be \eqn{n\times(n-1)/2}. We only need the upper
triangular matrix because the Hi-C matrix are symmetrical.}

\item{expected}{Underline true counts of the simulated data. For real data analysis, just set it as NULL.It takes
three formats that is the same as scHiC.}

\item{windowsize}{The size of neighborhood region. A windowsize of w results in a (2w+1)*(2w+1) neighboring submatirx.}

\item{nbins}{Number of bins of the observed single cell HiC matrix.}

\item{lambda1}{Tuning parameter to facilitate feature selection and regularization.}

\item{lambda2}{Tuning parameter to penalize teh diagonal element of the parameter to eliminate the trivial solution of representing an expression level as a linear combination of itself.}

\item{initA}{The initialization of A. The elements of A represent the similarities between loci in the same cell.}

\item{initS}{The initialization of S. The elements of S represent the similarities between all single cells at the same position.}

\item{ncores}{Number of cores to use. Default is 1.}

\item{MAX_ITER}{Maximum iteration of the external circulation of SRS.}

\item{ABSTOL}{Absolute tolerance of the external circulation.}

\item{learning_rate}{A hyper parameter that controls the speed of adjusting the weights of the network with respect to the loss gradient.}

\item{epochs}{The number of the entire training set going through the entire network.}

\item{batch_size}{The number of examples that are fed to the algorithm at a time.}

\item{run_batch}{Whether to use batch or to set the number of all the samples as teh value of the batch size. Default is TRUE.}

\item{verbose}{Whether to output the value of metrics at the end of each epoch. Default is TRUE.}

\item{estimates.only}{If TRUE, than out the SRS imputed matrix. If FALSE, A list of information is outputted.}
}
\value{

}
\description{
SRS
Self representation smoothing of single cell Hi-C matrix.
}
\examples{
SRS(scHiC, windowsize=2, nbins=61, learning_rate = 0.0001,epochs = 100)
}
